day_five/flask_jwt_practice/requirements.txt
flask==3.1.3
pyjwt==2.12.1
passlib[bcrypt]==1.7.4
pydantic[email]==2.13.3
sqlalchemy==2.0.49
psycopg2-binary==2.9.12
python-dotenv==1.2.1
alembic==1.16.5
pytest==8.4.2
pytest-flask==1.3.0
